requirements.txt
openai
bark
python-dotenv
